+++ zombienet-tests/src/0001-verify-contract-errors.ts
@@ -15,7 +15,6 @@ import { assert, stringToHex } from 'dedot/utils';
 import * as flipperV4Raw from '../flipper_v4.json';
 import * as flipperV5Raw from '../flipper_v5.json';
 import { FlipperContractApi } from './contracts/flipper';
-import { deployFlipper } from './shared';
 
 export const run = async (_nodeName: any, networkInfo: any) => {
   await cryptoWaitReady();
@@ -88,4 +87,39 @@ export const run = async (_nodeName: any, networkInfo: any) => {
   const apiV2 = await DedotClient.new(new WsProvider(wsUri));
   await verifyContracts(apiV2, flipperV4);
   await verifyContracts(apiV2, flipperV5);
+};
+
+const deployFlipper = async (api: ISubstrateClient, flipper: ContractMetadata, salt: string) => {
+  const alicePair = new Keyring({ type: 'sr25519' }).addFromUri('//Alice');
+  const caller = alicePair.address;
+
+  const wasm = flipper.source.wasm!;
+  const deployer = new ContractDeployer<FlipperContractApi>(api, flipper, wasm);
+
+  // Dry-run to estimate gas fee
+  const {
+    raw: { gasRequired },
+  } = await deployer.query.new(true, {
+    caller,
+    salt,
+  });
+
+  const contractAddress: string = await new Promise(async (resolve) => {
+    await deployer.tx.new(true, { gasLimit: gasRequired, salt }).signAndSend(alicePair, async ({ status, events }) => {
+      console.log(`[${api.rpcVersion}] Transaction status:`, status.type);
+
+      if (status.type === 'Finalized') {
+        const instantiatedEvent = events
+          .map(({ event }) => event) // prettier-end-here
+          .find(api.events.contracts.Instantiated.is); // narrow down the type for type suggestions
+
+        assert(instantiatedEvent, 'Event Contracts.Instantiated should be available');
+
+        const contractAddress = instantiatedEvent.palletEvent.data.contract.address();
+        resolve(contractAddress);
+      }
+    });
+  });
+
+  return contractAddress;
 };
